feat: Update BLE configuration and enhance debug flags
Updates the encryption keys for BLE communication and modifies MAC addresses for wheel devices. Enhances the debug system by extending debug flags to 16 bits, allowing additional flags for Bluetooth authentication and pairing events. Adds support for RFCOMM callback registration to improve the management of Bluetooth connections.

+++ esp32/arduino/remote_control/m25_ble.h
@@ -45,7 +45,7 @@
 #include <stddef.h>         // offsetof()
 
 // External debug flags (defined in serial_commands.h)
-extern volatile uint8_t debugFlags;
+extern volatile uint16_t debugFlags;
 
 // ---------------------------------------------------------------------------
 // M25 SPP Service / Characteristic UUIDs (m25_bluetooth.py)
